Update configurations for cc-mode
Delete part that set company-c-headers-path-system to be system
include paths defined in ede project because some system library are
too huge for cedet to process. All include paths are configured in
cc-mode-header-custom.el. cc-mode-header-custom.el and
cc-mode-projects.el are used to configure cc-mode projects and
examples are in utility folder.